Fix orphan (single-word last line) tappability by using line break strategy (#44)
Fixes #42.

Fixes tappability of regions relocated to a new line to avoid single-word last lines by UILabel. Xcode 12 (and the iOS 14 SDK) expose `.lineBreakStrategy` which is a new value on `UILabel` and `NSParagraphStyle` to control the orphan avoidance and line break positioning in strings.

This new method/enum is backwards compatible as it seems to be formerly-private API made public, so it works on all versions of iOS the label supports as long as you compile using Xcode 12.